+++ apps/client/components/ButtonArea/index.tsx
